Using long int as ModPow arguments type.

+++ lib/include/math_method.h
@@ -2,8 +2,8 @@
 
 namespace Salad {
 
-unsigned long ModPowNaive(unsigned base, unsigned exponent, unsigned modulus);
-unsigned long ModPow(unsigned base, unsigned exponent, unsigned modulus);
+unsigned long ModPowNaive(unsigned long base, unsigned long exponent, unsigned long modulus);
+unsigned long ModPow(unsigned long base, unsigned long exponent, unsigned long modulus);
 
 }
 // ~~ Salad::math_method
