+++ lib/src/math_method.cpp
@@ -4,12 +4,12 @@
 
 namespace Salad {
 
-unsigned long ModPowNaive(unsigned base, unsigned exponent, unsigned modulus)
+unsigned long ModPowNaive(unsigned long base, unsigned long exponent, unsigned long modulus)
 {
   return static_cast<unsigned long>(pow(base, exponent)) % modulus;
 }
 
-unsigned long ModPow(unsigned base, unsigned exponent, unsigned modulus)
+unsigned long ModPow(unsigned long base, unsigned long exponent, unsigned long modulus)
 {
   unsigned long result = 1;
   for (unsigned idx = 0; idx < exponent; ++idx) {
